add "proper" execute error handling
If sandbox-propagate-exceptions is set to false (the current default),
then exceptions are just caught in the evaluator and printed to stderr,
so Jupypter is not notified of the error per se.

With these changes (and if sandbox-propagate-exceptions is set to false),
an exception from evaluation is now communicated to Jupyter using an error
iopub message and an "error" execution_reply message.

As far as I can tell, it's easier to see errors printed to stderr than errors
reported to Jupyter, so I'm leaving propagate-exceptions off for now.